EclipsedSanity Alpha
structured room with dim lightning, assets for everything such as the Asylum, torches, doors, keys, weapons etc. An interact system with objects by pressing E on items. Key system for only allowing the player to open doors if they have the correct key, among with other features, will expand on this for the future.